fix: remove form actor even if not saved
fix #862
