+++ app/views/admin/job_offers/_form_actors.html.haml
@@ -25,6 +25,8 @@
                         = select_tag :email, options, include_blank: true, class: 'form-control', id: "email_#{actor_role}"
                       - else
                         = email_field_tag :email, nil, placeholder: t('simple_form.labels.defaults.email'), class: 'form-control', id: "email_#{actor_role}"
-                    = link_to('#', class: 'btn btn-primary btn-raised mb-0 d-flex flex-row align-items-center add_fields', data: {url: add_actor_admin_job_offers_url(role: actor_role, job_offer_id: job_offer_id)}) do
+                    - url = add_actor_admin_job_offers_url(role: actor_role, job_offer_id: job_offer_id)
+                    - data = { action: "job-offer-management#addFields", url: url}
+                    = link_to('#', class: 'btn btn-primary btn-raised mb-0 d-flex flex-row align-items-center', data: data) do
                       = t('buttons.add')
                       = fa_icon "plus", class: 'ml-1'
